Move const, fix extraData check, add extraData to tag and users
