+++ src/models/Idea.js
@@ -13,6 +13,8 @@ const merge = require('merge');
 
 var argVoteThreshold = config.ideas && config.ideas.argumentVoteThreshold;
 const userHasRole = require('../lib/sequelize-authorization/lib/hasRole');
+const getExtraDataConfig = require('../lib/sequelize-authorization/lib/getExtraDataConfig');
+
 
 module.exports = function (db, sequelize, DataTypes) {
 
@@ -217,73 +219,7 @@ module.exports = function (db, sequelize, DataTypes) {
       allowNull: true
     },
 
-    extraData: {
-      type: DataTypes.TEXT,
-      allowNull: false,
-      defaultValue: '{}',
-      get: function () {
-        let value = this.getDataValue('extraData');
-        try {
-          if (typeof value == 'string') {
-            value = JSON.parse(value);
-          }
-        } catch (err) {
-        }
-        return value;
-      },
-      set: function (value) {
-
-        try {
-          if (typeof value == 'string') {
-            value = JSON.parse(value);
-          }
-        } catch (err) {
-        }
-
-        let oldValue = this.getDataValue('extraData');
-        try {
-          if (typeof oldValue == 'string') {
-            oldValue = JSON.parse(oldValue) || {};
-          }
-        } catch (err) {
-        }
-
-        function fillValue(old, val) {
-          old = old || {};
-          Object.keys(old).forEach((key) => {
-            if (val[key] && typeof val[key] == 'object') {
-              return fillValue(old[key], val[key]);
-            }
-            if (val[key] === null) {
-              // send null to delete fields
-              delete val[key];
-            } else if (typeof val[key] == 'undefined') {
-              // not defined in put data; use old val
-              val[key] = old[key];
-            }
-          });
-        }
-
-        fillValue(oldValue, value);
-
-        this.setDataValue('extraData', JSON.stringify(value));
-      },
-      auth: {
-        authorizeData: function(self, action, user, data) {
-          if (!self.site) return; // todo: die kun je ophalen als eea. async is
-          data = data || self.extraData;
-          let result = {};
-          Object.keys(data).forEach((key) => {
-            let testRole = self.site.config && self.site.config.ideas && self.site.config.ideas.extraData[key] && self.site.config.ideas.extraData[key].auth && self.site.config.ideas.extraData[key].auth[action+'ableBy'];
-            testRole = testRole || ( self.auth && self.auth[action+'ableBy'] );
-            if (userHasRole(user, testRole, self.userId)) {
-              result[key] = data[key];
-            }
-          });
-          return result;
-        },
-      }
-    },
+    extraData: getExtraDataConfig(DataTypes.JSON,  'ideas'),
 
     location: {
       type: DataTypes.GEOMETRY('POINT'),
@@ -1330,7 +1266,7 @@ module.exports = function (db, sequelize, DataTypes) {
       return data;
     },
   }
-  
+
   return Idea;
 
   function beforeValidateHook(instance, options) {
