+++ src/routes/api/index.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const bruteForce = require('../../middleware/brute-force');
 
-let router = express.Router({mergeParams: true});
+const router = express.Router({mergeParams: true});
 
 // brute force
 router.use( bruteForce.globalMiddleware );
@@ -46,6 +46,7 @@ router.use( '/area', require('./area') );
 
 // output error as JSON only use this error handler middleware in "/api" based routes
 router.use("/site", function(err, req, res, next){
+  console.log('->>> err', err);
   // use the error's status or default to 500
   res.status(err.status || 500);
